Main

Description

Main Functionality is as Follows: - Contains a main thread which creates ten “users” (ten values of type User), and spawns ten threads, one for each of these users. - Each user thread should behave as follows: at random time intervals, the thread should select one of the other users at random, and send a random message to that user. - Simulate 100 messages, and then terminate and output the final count of how many messages each user received.

Additional Feature: - Message history of a user; includes sent and received messages.

Documentation

main :: IO ()